+++ packages/transducers-stats/src/index.ts
@@ -2,13 +2,16 @@ export * from "./bollinger";
 export * from "./donchian";
 export * from "./ema";
 export * from "./hma";
+export * from "./macd";
 export * from "./momentum";
 export * from "./roc";
 export * from "./rsi";
 export * from "./sd";
 export * from "./sma";
+export * from "./stochastic";
 export * from "./trix";
 export * from "./wma";
 
+export * from "./bounds";
 export * from "./dot";
 export * from "./mse";