+++ packages/transducers-stats/src/wma.ts
@@ -11,8 +11,7 @@ import { dot } from "./dot";
  * https://en.wikipedia.org/wiki/Moving_average#Weighted_moving_average
  *
  * Note: the number of results will be `period-1` less than the number
- * of processed inputs and no outputs will be produced if there were
- * less than `period` input values.
+ * of processed inputs.
  *
  * @param weights period or array of weights
  */