+++ packages/iges/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/thi-ng/umbrella/compare/@thi.ng/iges@0.1.3...@thi.ng/iges@0.1.4) (2018-07-25)
+
+
+
+
+**Note:** Version bump only for package @thi.ng/iges
+
 <a name="0.1.3"></a>
 ## [0.1.3](https://github.com/thi-ng/umbrella/compare/@thi.ng/iges@0.1.2...@thi.ng/iges@0.1.3) (2018-07-20)